requirements.txt
flask
flask-bcrypt
flask-wtf
email-validator
flask-login
Pillow
itsdangerous
flask-mail
